mayannah/requirements.txt
arrow==0.10.0
certifi==2017.4.17
chardet==3.0.4
contextlib2==0.5.5
coolname==0.2.1
defusedxml==0.5.0
Django==1.11.2
django-allauth==0.32.0
django-debug-toolbar==1.8
django-environ==0.4.3
djangorestframework==3.6.3
djangorestframework-jwt==1.11.0
gunicorn==19.7.1
idna==2.5
oauthlib==2.0.2
psycopg2==2.7.1
python-dateutil==2.6.0
python3-openid==3.1.0
pytz==2017.2
raven==6.1.0
requests==2.18.1
requests-oauthlib==0.8.0
six==1.10.0
sqlparse==0.2.3
urllib3==1.21.1
whitenoise==3.3.0
